dofi plugin: extended command server to cope with 64bit return values several improvements of dofi gui, use table of and/or bits

+++ plugins/dofi/dofi/TerminalModule.h
@@ -184,7 +184,7 @@ typedef enum
   //~ int fErrcount; /**< errorcount for verify*/
 
 	/*** keep result of most recent command call here */
-	std::vector<long> fCommandResults;
+	std::vector<unsigned long long> fCommandResults;
 
 		 /*** keep address of most recent command call here*/
 	std::vector<long> fCommandAddress;
